dom/indexedDB/test/test_file_cross_database_copying.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript">
  function* testSteps()
  {
    const READ_WRITE = "readwrite";

    const databaseInfo = [
      { name: window.location.pathname + "1" },
      { name: window.location.pathname + "2" }
    ];

    const objectStoreName = "Blobs";

    const fileData = { key: 1, file: getRandomFile("random.bin", 100000) };

    SpecialPowers.pushPrefEnv({ set: [["dom.indexedDB.dataThreshold", -1]] },
                              continueToNextStep);
    yield undefined;

    let databases = [];
    for (let info of databaseInfo) {
      let request = indexedDB.open(info.name, 1);
      request.onerror = errorHandler;
      request.onupgradeneeded = grabEventAndContinueHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield undefined;

      is(event.type, "upgradeneeded", "Got correct event type");

      let db = event.target.result;
      db.onerror = errorHandler;

      let objectStore = db.createObjectStore(objectStoreName, { });
      objectStore.add(fileData.file, fileData.key);

      event = yield undefined;

      is(event.type, "success", "Got correct event type");

      databases.push(db);
    }

    let refResult;
    for (let db of databases) {
      let request = db.transaction([objectStoreName])
                      .objectStore(objectStoreName).get(fileData.key);
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield undefined;

      let result = event.target.result;
      verifyBlob(result, fileData.file, 1);
      yield undefined;

      if (!refResult) {
        refResult = result;
        continue;
      }

      isnot(getFilePath(result), getFilePath(refResult), "Different os files");
    }

    for (let i = 1; i < databases.length; i++) {
      let db = databases[i];

      let objectStore = db.transaction([objectStoreName], READ_WRITE)
                          .objectStore(objectStoreName);

      let request = objectStore.add(refResult, 2);
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield undefined;

      is(event.target.result, 2, "Got correct key");

      request = objectStore.get(2);
      request.onsuccess = grabEventAndContinueHandler;
      event = yield undefined;

      let result = event.target.result;
      verifyBlob(result, refResult, 2);
      yield undefined;

      isnot(getFilePath(result), getFilePath(refResult), "Different os files");
    }

    is(bufferCache.length, 2, "Correct length");

    finishTest();
  }
  </script>
  <script type="text/javascript" src="file.js"></script>
  <script type="text/javascript" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
